MapEditor 2D View
Updated the MapEditor's 2D view to now be able to paint tiles. Still working to get the 3D view started, but the 2D view is now centered and the tile palette, currently the global tile database, is moved to the left. The tools are being planned to be above the either the tile palette on the left, or above the 2D view in the center, still unsure, but will continue to update.